+++ 1.SemesterProjekt/Models/Product.cs
@@ -16,7 +16,7 @@ namespace _1.SemesterProjekt.Models
         /// <param name="brand"></param>
         /// <param name="price"></param>
         /// <param name="stock"></param>
-        public Product(int id, string name, string brand, double price, int stock) 
+        public Product(int id, string name, string brand, decimal price, int stock) 
         { 
             this.ID = id;
             this.Name = name;
@@ -28,7 +28,7 @@ namespace _1.SemesterProjekt.Models
         public int ID { get; set; }
         public string Name { get; set; }
         public string Brand { get; set; }
-        public double Price { get; set; }
+        public decimal Price { get; set; }
         public int Stock { get; set; }
 
     }
